+++ components/Divider.tsx
@@ -1,6 +1,10 @@
 import React from "react";
 
-function Divider({ color }) {
+interface DividerProps {
+  color?: string
+}
+
+function Divider({ color }: DividerProps) {
   return <div className={`w-[90%] h-[1px] bg-${color ?? "offWhite"} my-5`} />;
 }
