+++ components/profile/OfficeAddressForm.tsx
@@ -3,64 +3,41 @@ import { ErrorMessage, Field, Form, Formik } from "formik";
 import * as Yup from "yup";
 import { useDispatch } from "react-redux";
 import { useAppSelector } from "@/store/hooks";
-import {
-  selectBillingAddress,
-  updateBillingAddress,
-} from "@/store/userSlice";
+import { selectOfficeAddress, updateOfficeAddress } from "@/store/userSlice";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPenToSquare } from "@fortawesome/free-regular-svg-icons";
-import { ToastContainer } from "react-toastify";
-import {
-  faCircleDollarToSlot,
-  faLocationDot,
-} from "@fortawesome/free-solid-svg-icons";
+import { ToastContainer, toast } from "react-toastify";
+import { faLocationDot, faTruck } from "@fortawesome/free-solid-svg-icons";
 import { useTranslation } from "react-i18next";
 import AddInfoButton from "./AddInfoButton";
+import { FormProps } from "@/models/forms";
 
-function BillingAddressForm({ disabled, setShowError }) {
+function OfficeAddressForm({ disabled, setShowError }: FormProps) {
   const dispatch = useDispatch();
   const { t } = useTranslation(["profile", "buttons", "forms"]);
-  const billingAddressInfo = useAppSelector(selectBillingAddress);
+  const officeAddressInfo = useAppSelector(selectOfficeAddress);
   const [isPressed, setIsPressed] = useState(false);
-
   const [shouldShowAddress, setShouldShowAddress] = useState(
-    billingAddressInfo ? true : false
+    officeAddressInfo ? true : false
   );
 
   const labels = {
-    company: t("company", { ns: "forms" }),
-    street: t("street", { ns: "forms" }),
-    streetNumber: t("streetNumber", { ns: "forms" }),
-    houseNumber: t("houseNumber", { ns: "forms" }),
-    city: t("city", { ns: "forms" }),
-    postalCode: t("postalCode", { ns: "forms" }),
+    officeAddress: t("officeAddress", { ns: "forms" }),
   };
 
   const initialValues = {
-    city: billingAddressInfo?.city ?? "",
-    company: billingAddressInfo?.company ?? "",
-    street: billingAddressInfo?.street ?? "",
-    streetNumber: billingAddressInfo?.streetNumber ?? "",
-    houseNumber: billingAddressInfo?.houseNumber ?? "",
-    postalCode: billingAddressInfo?.postalCode ?? "",
+    officeAddress: officeAddressInfo?.officeAddress ?? "",
   };
 
   const [address, setAddress] = useState({ ...initialValues });
 
   const isEmptyAddress = Object.values(address).every((k) => k.length < 1);
-  const addressSaved = !!billingAddressInfo;
+  const addressSaved = !!officeAddressInfo;
 
   const validationSchema = Yup.object().shape({
-    company: Yup.string().optional(),
-    city: Yup.string()
-      .required(t("city_required", { ns: "forms" }))
-      .min(4, t("city_min", { ns: "forms" })),
-    street: Yup.string().required(t("street_required", { ns: "forms" })),
-    streetNumber: Yup.string().required(
-      t("streetNumber_required", { ns: "forms" })
-    ),
-    houseNumber: Yup.string().required(t("house_required", { ns: "forms" })),
-    postalCode: Yup.string().required(t("postal_required", { ns: "forms" })),
+    officeAddress: Yup.string()
+      .required(t("officeAddress_required", { ns: "forms" }))
+      .min(10, t("officeAddress_min", { ns: "forms" })),
   });
 
   //=========================== Handler Functions START ============================//
@@ -70,11 +47,10 @@ function BillingAddressForm({ disabled, setShowError }) {
   };
 
   const handleSubmit = async () => {
-    dispatch(updateBillingAddress(address));
-
+    dispatch(updateOfficeAddress(address));
     setShouldShowAddress(true);
 
-    setShowError(false);
+    setShowError?.(false);
 
     toast.success(t("address_saved", { ns: "buttons" }), {
       position: "bottom-right",
@@ -94,16 +70,16 @@ function BillingAddressForm({ disabled, setShowError }) {
 
   if (shouldShowAddress)
     return (
-      <BillingAddressInfo
+      <OfficeAddressInfo
         disabled={disabled}
-        addressInfo={billingAddressInfo}
+        officeAddressInfo={officeAddressInfo}
         setShouldShowAddress={setShouldShowAddress}
       />
     );
 
-  if (!shouldShowAddress && !isPressed && !billingAddressInfo) {
-    return <AddInfoButton label={"billing_address"} onPress={() => setIsPressed(!isPressed)}/>
-  }
+    if (!shouldShowAddress && !isPressed && !officeAddressInfo) {
+      return <AddInfoButton label={"officeAddress"} onPress={() => setIsPressed(!isPressed)}/>
+    }
 
   return (
     <div className="bg-gradient-to-r from-[#0ba360] to-[#3cba92] flex-1 flex-column space-y-2 p-[40px] shadow-md border-zinc-700 rounded-sm">
@@ -113,7 +89,7 @@ function BillingAddressForm({ disabled, setShowError }) {
           <FontAwesomeIcon color="#EEEEEE" icon={faLocationDot} />
         </div>
         <h2 className="font-semibold text-xl text-[#F8F1F1]">
-          {t("billing_address", { ns: "forms" })}
+          {t("officeAddress", { ns: "forms" })}
         </h2>
       </div>
       <Formik
@@ -151,21 +127,21 @@ function BillingAddressForm({ disabled, setShowError }) {
                 <div className="flex flex-row gap-5">
                   <button
                     onClick={() =>
-                      addressSaved ? setShouldShowAddress(true) : setIsPressed(false)
+                      addressSaved ? setShouldShowAddress(true) : setIsPressed(!isPressed)
                     }
                     className={`
-                     hover:opacity-80 
-                     transition 
-                     ease-in-out
-                     md:mr-auto 
-                     md:ml-0
-                     min-h-[50px] 
-                    bg-[#181516] 
-                     rounded-[4px] 
-                     md:max-w-[450px] 
-                     w-full 
-                    text-offWhite
-                    `}
+                   hover:opacity-80
+                   transition 
+                   ease-in-out"
+                   md:mr-auto 
+                   md:ml-0 
+                   min-h-[50px] 
+                   bg-[#181516] 
+                   rounded-[4px] 
+                   md:max-w-[450px] 
+                   w-full 
+                   text-offWhite
+                   `}
                     type="button"
                   >
                     {t("cancel", { ns: "buttons" })}
@@ -194,9 +170,9 @@ function BillingAddressForm({ disabled, setShowError }) {
   );
 }
 
-export const BillingAddressInfo = ({
+export const OfficeAddressInfo = ({
   disabled,
-  addressInfo,
+  officeAddressInfo,
   setShouldShowAddress,
 }) => {
   const { t } = useTranslation(["profile", "buttons"]);
@@ -204,10 +180,10 @@ export const BillingAddressInfo = ({
     <div>
       <div className="flex items-center gap-2 mb-2">
         <div className="bg-gradient-to-r from-[#0ba360] to-[#3cba92] w-8 h-8 rounded-full flex items-center justify-center">
-          <FontAwesomeIcon color="#EEEEEE" icon={faCircleDollarToSlot} />
+          <FontAwesomeIcon color="#EEEEEE" icon={faTruck} />
         </div>
         <h2 className="text-offWhite text-xl font-semibold">
-          {t("billing_address", { ns: "forms" })}
+          {t("officeAddress", { ns: "forms" })}
         </h2>
       </div>
       <div className="relative">
@@ -234,18 +210,17 @@ export const BillingAddressInfo = ({
         <div
           className={`${
             disabled
-              ? "opacity-60 pointer-events-none border-[0px]"
+              ? "opacity-60 pointer-events-none"
               : "opacity-100 border-[2px] border-neonGreen"
-          } bg-offWhite p-4 rounded-md flex flex-col gap-1 transition ease-in-out`}
+          } bg-offWhite p-4 rounded-md flex flex-col gap-1`}
         >
-          {addressInfo?.company && <div>{`${addressInfo.company}`}</div>}
-          <div className="text-[#151718]">{addressInfo.city}</div>
-          <div className="text-[#151718]">{`${addressInfo.postalCode}`}</div>
-          <div className="text-[#151718]">{`${addressInfo.street} ${addressInfo.streetNumber}`}</div>
+          <div className="text-[#151718]">
+            {officeAddressInfo.officeAddress}
+          </div>
         </div>
       </div>
     </div>
   );
 };
 
-export default BillingAddressForm;
+export default OfficeAddressForm;
