+++ models/forms.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface FormProps {
+  disabled?: boolean;
+  setShowError?: Dispatch<SetStateAction<boolean>>;
+}
