+++ package.json
@@ -22,6 +22,8 @@
         "@stripe/react-stripe-js": "^1.16.5",
         "@stripe/stripe-js": "^1.48.0",
         "@tailwindcss/typography": "^0.5.10",
+        "@types/react": "^18.2.37",
+        "@types/react-toastify": "^4.1.0",
         "@vercel/analytics": "^1.1.1",
         "axios": "^1.4.0",
         "babel-plugin-macros": "^3.1.0",
@@ -51,12 +53,15 @@
         "stripe": "^13.4.0",
         "swr": "^2.1.0",
         "tailwindcss-filters": "^3.0.0",
+        "touch-cli": "^0.0.1",
         "yup": "^1.2.0"
     },
     "devDependencies": {
         "@babel/preset-react": "^7.22.5",
+        "@types/node": "^20.9.0",
         "autoprefixer": "^10.4.14",
         "postcss": "^8.4.21",
-        "tailwindcss": "^3.2.7"
+        "tailwindcss": "^3.2.7",
+        "typescript": "^5.2.2"
     }
 }
