+++ pages/forgot_password.tsx
@@ -51,6 +51,7 @@ const ForgotPassword = () => {
   return (
     <div className="min-h-[650px] flex items-center">
       <Wrapper>
+        <>
         <ToastContainer />
         <div className="max-w-[600px] rounded-lg p-5 border text-offWhite border-offWhite mx-auto flex flex-col">
           <div className="text-2xl font-bold text-offWhite">
@@ -110,6 +111,7 @@ const ForgotPassword = () => {
             )}
           </Formik>
         </div>
+        </>
       </Wrapper>
     </div>
   );
