+++ utils/toastConfig.ts
@@ -1,15 +1,6 @@
-export const errorConfig = {
-  position: "bottom-right",
-  autoClose: 5000,
-  hideProgressBar: false,
-  closeOnClick: true,
-  pauseOnHover: true,
-  draggable: true,
-  progress: undefined,
-  theme: "dark",
-};
+import { ToastOptions } from "react-toastify"
 
-export const successConfig = {
+export const errorConfig: ToastOptions = {
   position: "bottom-right",
   autoClose: 5000,
   hideProgressBar: false,
@@ -18,4 +9,15 @@ export const successConfig = {
   draggable: true,
   progress: undefined,
   theme: "dark",
-};
+} 
+
+export const successConfig: ToastOptions = {
+  position: "bottom-right",
+  autoClose: 5000,
+  hideProgressBar: false,
+  closeOnClick: true,
+  pauseOnHover: true,
+  draggable: true,
+  progress: undefined,
+  theme: "dark",
+}
